+++ src/main/java/nuparu/sevendaystomine/entity/ZombiePigEntity.java
@@ -11,10 +11,7 @@ import net.minecraft.util.DamageSource;
 import net.minecraft.util.SoundEvent;
 import net.minecraft.util.SoundEvents;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.MathHelper;
 import net.minecraft.world.World;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
 import nuparu.sevendaystomine.init.ModEntities;
 
 public class ZombiePigEntity<T extends ZombiePigEntity> extends ZombieQuadrapedEntity {
@@ -35,7 +32,7 @@ public class ZombiePigEntity<T extends ZombiePigEntity> extends ZombieQuadrapedE
 	
 	public static AttributeModifierMap createAttributes() {
 		return MonsterEntity.createMonsterAttributes( ).add(Attributes.FOLLOW_RANGE, 32.0D)
-				.add(Attributes.MOVEMENT_SPEED, (double) 0.19f).add(Attributes.ATTACK_DAMAGE, 3.0D)
+				.add(Attributes.MOVEMENT_SPEED, 0.19f).add(Attributes.ATTACK_DAMAGE, 3.0D)
 				.add(Attributes.ARMOR, 0.0D).add(Attributes.MAX_HEALTH, 50).build();
 	}
